requirements.txt
langchain
pypdf
streamlit
boto3
langchain-community
langchain-aws
awswrangler
pymongo
ipykernel
openai
redshift_connector
python-dotenv